api/requirements.txt
fastapi
uvicorn[standard]
pymongo
neo4j
cassandra-driver
python-dotenv>=1.0
